refactor(app): extract waitForWorktree from submit.ts
Move the per-submit waitForWorktree closure out of submit.ts into a
createWaitForWorktree factory. It receives restoreInput as an injected
bound VoidFunction, so the new module has no SubmitOwnership or draft
lifecycle knowledge. The shared pending registry and WorktreeState are
imported directly; submit.ts still uses both outside this closure.

Pure extraction: no behavior, DOM, aria, copy, or storage-key change.
